framework/tests/test_error_handler.py
```python
import unittest
from framework.error_handler import ErrorHandler

class TestErrorHandler(unittest.TestCase):

    def setUp(self):
        self.error_handler = ErrorHandler()

    def test_log_error(self):
        try:
            raise Exception("Test Exception")
        except Exception as e:
            self.error_handler.log_error(e)
            self.assertTrue(self.error_handler.error_log)

    def test_alert(self):
        try:
            raise Exception("Test Exception")
        except Exception as e:
            alert_message = self.error_handler.alert(e)
            self.assertEqual(alert_message, "An error occurred: Test Exception")

    def test_clear_log(self):
        try:
            raise Exception("Test Exception")
        except Exception as e:
            self.error_handler.log_error(e)
            self.error_handler.clear_log()
            self.assertFalse(self.error_handler.error_log)

if __name__ == '__main__':
    unittest.main()
```